# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Backend API contract for the Nigeria Security mobile app.
servers:
  - url: /api
    description: Base API path
...
security:
  - bearerAuth: []
paths:
...
  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List reports visible to the current user
      parameters:
        - $ref: "#/components/parameters/AgencyQuery"
        - $ref: "#/components/parameters/StatusQuery"
        - $ref: "#/components/parameters/EmergencyQuery"
        - $ref: "#/components/parameters/SourceQuery"
        - $ref: "#/components/parameters/SearchQuery"
      responses:
        "200":
          description: Report list
          content:
            application/json:
              schema:
                type: object
                properties:
                  reports:
                    type: array
                    items:
                      $ref: "#/components/schemas/ReportSummary"
                required: [reports]
        "401":
          $ref: "#/components/responses/Unauthorized"
  /agencies/{agency}/reports:
    get:
      operationId: listAgencyReports
      tags: [reports]
      summary: List reports routed to one agency
      parameters:
        - $ref: "#/components/parameters/AgencyPath"
        - $ref: "#/components/parameters/StatusQuery"
        - $ref: "#/components/parameters/EmergencyQuery"
        - $ref: "#/components/parameters/SourceQuery"
      responses:
        "200":
          description: Agency report list
          content:
            application/json:
              schema:
                type: object
                properties:
                  reports:
                    type: array
                    items:
                      $ref: "#/components/schemas/ReportSummary"
                required: [reports]
        "403":
          $ref: "#/components/responses/Forbidden"
  /agencies/{agency}/dashboard:
    get:
      operationId: getAgencyDashboardMetrics
      tags: [reports]
      summary: Get agency dashboard metrics
      parameters:
        - $ref: "#/components/parameters/AgencyPath"
      responses:
        "200":
          description: Agency dashboard metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgencyDashboardMetrics"
        "403":
          $ref: "#/components/responses/Forbidden"
  /reports/{reportId}:
    get:
      operationId: getReportDetail
      tags: [reports]
      summary: Get report detail
      parameters:
        - $ref: "#/components/parameters/ReportIdParam"
      responses:
        "200":
          description: Report detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportDetail"
        "404":
          $ref: "#/components/responses/NotFound"
  /reports/{reportId}/status:
    patch:
      operationId: updateReportStatus
      tags: [reports]
      summary: Update report status
      parameters:
        - $ref: "#/components/parameters/ReportIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportStatusUpdateRequest"
      responses:
        "200":
          description: Updated report detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportDetail"
        "403":
          $ref: "#/components/responses/Forbidden"
  /reports/{reportId}/reassign:
    post:
      operationId: reassignReport
      tags: [reports]
      summary: Reassign a report to another agency
      description: Admin-only operation for changing agency ownership/routing.
      parameters:
        - $ref: "#/components/parameters/ReportIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportReassignmentRequest"
      responses:
        "200":
          description: Reassigned report detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportDetail"
        "403":
          $ref: "#/components/responses/Forbidden"
  /reports/{reportId}/timeline:
    get:
      operationId: listReportTimeline
      tags: [reports]
      summary: List report timeline/history
      parameters:
        - $ref: "#/components/parameters/ReportIdParam"
      responses:
        "200":
          description: Timeline entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  timeline:
                    type: array
                    items:
                      $ref: "#/components/schemas/TimelineEntry"
                required: [timeline]
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    AgencyPath:
      name: agency
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/AgencyType"
    AgencyQuery:
      name: agency
      in: query
      schema:
        $ref: "#/components/schemas/AgencyType"
    StatusQuery:
      name: status
      in: query
      schema:
        $ref: "#/components/schemas/ReportStatus"
    EmergencyQuery:
      name: emergencyLevel
      in: query
      schema:
        $ref: "#/components/schemas/EmergencyLevel"
    SourceQuery:
      name: source
      in: query
      schema:
        $ref: "#/components/schemas/ReportSource"
    SearchQuery:
      name: q
      in: query
      schema:
        type: string
...
    ReportIdParam:
      name: reportId
      in: path
      required: true
      schema:
        type: string
        minLength: 3
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    AgencyDashboardMetrics:
      type: object
      properties:
        agency:
          type: string
        metrics:
          type: object
          properties:
            total: { type: integer }
            submitted: { type: integer }
            triaged: { type: integer }
            assigned: { type: integer }
            in_progress: { type: integer }
            resolved: { type: integer }
            closed: { type: integer }
            rejected: { type: integer }
            highPriority: { type: integer }
            withCoordinates: { type: integer }
            withoutCoordinates: { type: integer }
          required:
            [
              total,
              submitted,
              triaged,
              assigned,
              in_progress,
              resolved,
              closed,
              rejected,
              highPriority,
              withCoordinates,
              withoutCoordinates,
            ]
      required: [agency, metrics]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        issues:
          type: object
          additionalProperties: true
      required: [error]
    AgencyType:
      type: string
      enum: [frsc, police, vio, civil_defence, admin, citizen, dss, fire_service, custom]
...
    ReportStatus:
      type: string
      enum: [submitted, triaged, assigned, in_progress, resolved, closed, rejected]
    EmergencyLevel:
      type: string
      enum: [low, medium, high, critical]
    ReportSource:
      type: string
      enum: [citizen_report, agency_referral, officer_created, admin_created]
    IncidentType:
      type: string
      enum:
        - road_crash
        - traffic_obstruction
        - dangerous_driving
        - vehicle_breakdown
        - road_hazard
        - theft
        - robbery
        - assault
        - suspicious_activity
        - public_threat
        - vehicle_theft
        - missing_vehicle_alert
        - security_incident
        - roadworthiness_complaint
        - dangerous_vehicle
        - vehicle_inspection_concern
        - invalid_certificate_concern
        - commercial_vehicle_safety_issue
        - vehicle_documentation_concern
        - fire_rescue
        - disaster
        - civil_emergency
        - infrastructure_risk
        - custom
    LocationSource:
      type: string
      enum: [gps, manual]
    Location:
      type: object
      properties:
        latitude:
          type: number
          minimum: -90
          maximum: 90
        longitude:
          type: number
          minimum: -180
          maximum: 180
        address:
          type: string
        state:
          type: string
        lga:
          type: string
        source:
          $ref: "#/components/schemas/LocationSource"
        accuracy:
          type: number
      required: [address, source]
    Reporter:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        email:
          type: string
          format: email
      required: [phone]
    VehicleInfo:
      type: object
      properties:
        registrationNumber:
          type: string
        make:
          type: string
        model:
          type: string
        color:
          type: string
...
    ReportSummary:
      type: object
      properties:
        id:
          type: string
        reference:
          type: string
        incidentType:
          $ref: "#/components/schemas/IncidentType"
        description:
          type: string
        suggestedAgency:
          $ref: "#/components/schemas/AgencyType"
        assignedAgency:
          $ref: "#/components/schemas/AgencyType"
        emergencyLevel:
          $ref: "#/components/schemas/EmergencyLevel"
        status:
          $ref: "#/components/schemas/ReportStatus"
        source:
          $ref: "#/components/schemas/ReportSource"
        location:
          $ref: "#/components/schemas/Location"
        submittedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, reference, incidentType, suggestedAgency, emergencyLevel, status, source, submittedAt, updatedAt]
    ReportDetail:
      allOf:
        - $ref: "#/components/schemas/ReportSummary"
        - type: object
          properties:
            reporter:
              $ref: "#/components/schemas/Reporter"
            vehicle:
              $ref: "#/components/schemas/VehicleInfo"
            evidence:
              type: array
              items:
                $ref: "#/components/schemas/EvidenceMetadata"
            timeline:
              type: array
              items:
                $ref: "#/components/schemas/TimelineEntry"
            citizenMessage:
              type: string
          required: [timeline]
...
    ReportStatusUpdateRequest:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/ReportStatus"
        note:
          type: string
        assignedToUserId:
          type: string
      required: [status]
    ReportReassignmentRequest:
      type: object
      properties:
        targetAgency:
          $ref: "#/components/schemas/AgencyType"
        reason:
          type: string
        note:
          type: string
      required: [targetAgency, reason]
    TimelineEntry:
      type: object
      properties:
        id:
          type: string
        reportId:
          type: string
        status:
          $ref: "#/components/schemas/ReportStatus"
        action:
          type: string
        message:
          type: string
        actorUserId:
          type: string
        actorAgency:
          $ref: "#/components/schemas/AgencyType"
        createdAt:
          type: string
          format: date-time
      required: [id, action, message, createdAt]
    EvidenceKind:
      type: string
      enum: [photo, video, audio, document, statement, other]
    EvidenceCreateRequest:
      type: object
      properties:
        kind:
          $ref: "#/components/schemas/EvidenceKind"
        uri:
          type: string
        fileName:
          type: string
        mimeType:
          type: string
        sizeBytes:
          type: integer
        checksum:
          type: string
        capturedAt:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties: true
      required: [kind, uri]
    EvidenceMetadata:
      allOf:
        - $ref: "#/components/schemas/EvidenceCreateRequest"
        - type: object
          properties:
            id:
              type: string
            reportId:
              type: string
            uploadedByUserId:
              type: string
            createdAt:
              type: string
              format: date-time
          required: [id, reportId, createdAt]
...